clean up key pressing
