+++ src/canva/renderer/Renderer.tsx
@@ -39,6 +39,7 @@ import {
 import { d3Listeners } from "./core/d3Main";
 import { getCanvasState } from "./core/exposed/getCanvasState";
 import { actionClickHandler } from "./core/actionClickHandler";
+import { zoomShortcutHandler } from "./core/zoomShortcutHandler";
 
 interface ReactInfiniteCanvasRendererProps extends ReactInfiniteCanvasProps {
   children: any;
@@ -178,41 +179,10 @@ export const ReactInfiniteCanvasRenderer = memo(
       [fitContentToView]
     );
 
-    useEffect(
-      function zoomShortcutHandler() {
-        function onKeyDownHandler(e: KeyboardEvent) {
-          const cmdPressed = e.metaKey || e.ctrlKey;
-          if (!Object.values(ZOOM_KEY_CODES).includes(e.which) || !cmdPressed)
-            return;
-          e.preventDefault();
-          switch (e.which) {
-            case ZOOM_KEY_CODES.ZOOM_IN:
-            case ZOOM_KEY_CODES.ZOOM_IN_2:
-              onActionClick(ZOOM_CONTROLS.ZOOM_IN);
-              break;
-            case ZOOM_KEY_CODES.ZOOM_OUT:
-            case ZOOM_KEY_CODES.ZOOM_OUT_2:
-              onActionClick(ZOOM_CONTROLS.ZOOM_OUT);
-              break;
-            case ZOOM_KEY_CODES.FIT_TO_VIEW:
-              onActionClick(ZOOM_CONTROLS.FIT_TO_VIEW);
-              break;
-            case ZOOM_KEY_CODES.FIT_TO_HUNDRED:
-              onActionClick(ZOOM_CONTROLS.FIT_TO_HUNDRED);
-              break;
-            default:
-              return;
-          }
-        }
-
-        window.addEventListener("keydown", onKeyDownHandler);
-        return () => {
-          window.removeEventListener("keydown", onKeyDownHandler);
-        };
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-      },
-      [onActionClick]
-    );
+    useEffect(() => {
+      const cleanup = zoomShortcutHandler({ onActionClick });
+      return cleanup;
+    }, [onActionClick]);
 
     const getContainerOffset = useCallback(function offsetHandler(
       isVertical = true
